Cherry-pick Ie8cdd44de44cf68c15541f6b4fdb657e640a9d1a for ZP1A.250822.001
Fix build error in CarrierConfigLoader\n\nFixes a build error in CarrierConfigLoader.java caused by the removal of the `supportCarrierServicesForHsum` feature flag.\n\nThe build failed with a \"cannot find symbol\" error for the `supportCarrierServicesForHsum` method on the `FeatureFlags` object. This method was used in two places to conditionally call either a standard API or a newer `...AsUser` variant for multi-user environments.\n\nThe removal of the feature flag indicates that the `...AsUser` path should now be used by default. This patch removes the conditional checks and updates the code to directly use the `...AsUser` APIs in both locations.\n\nBug: 357638490

Bug: 441166917
Change-Id: Ic5a6a64f4d21dbeee9801d2c3ba8cdcdaf6aba57