TelephonyShellCommand to uncap max allowed satellite data mode
The Satellite Data Policy is cappped to an OEM config
max_allowed_data_mode. But for testing purposes, we have to override
this, so that data mode could be uncapped and we could test any
satellite data mode. Currenely there are couple of ways to achieve this
-- either through PhoneInfo or through ConfigUpdator, but they are not
really developer friendly options. So, introduce a shell command for the
same:
`adb shell cmd phone uncap-max-allowed-satellite-data-mode`

Bug: b/447209974
Flag: EXEMPT BUGFIX
Test: adb shell cmd phone uncap-max-allowed-satellite-data-mode
Change-Id: I9f2740ea892360204a4299bcbc3ab0ad7d34415d