TelecomAccountRegistry async broadcast logic
Move the broadcast logic in TelecomAccountRegistry to be executed async.
We have seen several ANR reports in
TelecomAccountRegistry#setupOnBootInternal on watch devices. This seems
to be due to the expensives operations we perform by tearing down and
re-registering phone accounts on the main thread when we receive
broadcasts. These should be done on a separate handler thread instead.

Bug: 451460260
Flag: EXEMPT BUGFIX
Test: atest TelecomAccountRegistryTest
Test: atest TelephonyConnectionServiceTest
Change-Id: I14e07bff39f2a3e68fbca61666b8cabafcc1d3fd