Clear mPersistentOverrideConfigs when clearing config for phone
Previously,  and  was not reset in . This resulted in a stale state where, if a SIM card with a persistent override was replaced by a SIM card without one, the old override configuration would remain in memory.

When a new SIM is loaded,  returns null if no override file exists, causing the code to skip updating  &  and leaving the previous value intact.

This CL explicitly clears  and  in  to ensure that configuration overrides do not leak between different SIMs.

Bug: 445555865
Flag: EXEMPT BUGFIX
Test: Manually Tested
Change-Id: I8817916590707e9461c45415ae69299889e345fe