Fix crash in PhoneInformationV2FragmentDataNetwork on service state change
The application crashes with an IllegalStateException when updateServiceState is called because the fragment might be detached from the context when the onServiceStateChanged callback is fired. Accessing resources via getResources() on a detached fragment causes this exception.

This change adds an isAdded() check to ensure the fragment is attached to a context before attempting to access resources or update the UI.

Bug: 464138800
Test: Manually Tested
Flag: EXEMPT BUGFIX
Change-Id: I1726a991a06e9f1b1724effaaf38c3e8d7facbb4